feat: 工具中的烧录抽离
